test backward pass in jade

+++ test_imagenet.py
@@ -315,7 +315,7 @@ def load_small_imagenet(args: dict, val_size=5000, test_size=10000, shuffle_val=
     #                                  std=[0.229, 0.224, 0.225])
 
     transform_test = transforms.Compose([
-        transforms.Resize((args["resolution"] * ratio)),
+        transforms.Resize(int(args["resolution"] * ratio)),
         transforms.CenterCrop(args["resolution"]),
         transforms.ToTensor(),
         transforms.Normalize([0.4824, 0.4495, 0.3981], [0.2301, 0.2264, 0.2261]),
